Remove Hard Dependencies on Hideous Helmet
This means I'll have to re-implement the logic to check for the helmet, or in the case of the Wound Counter, fully implement it; but seeing as how they sort of already function that way, this shouldn't be the worst.  It also just means keeping up with HH's updates to these respective HUD Elements and pulling in anything that affects their behavior.