Refactor: "I didn't push for a week" Edition
This started as an experiment to see how hard it would be to add units of measurement to the compass, but now has ballooned into a thought experiment for how to best implement dataclasses in zscript, if their primary purpose is information in -> information out + unique arithmetic. The compass itself has varying features planned, which I've reimplemented twice now, and am in the process of redoing again. Those features are as follows:
- Ability to convert MU coordinates into Meters, Kilometers, and Miles,
- Ability to display the cardinal directions the player is in relative to the origin coordinates of the map (x->East/West, y->North/South, z->Up/Down),
- Ability to display the X, Y, and Z labels for a position,
- Linebreaks in the positional string,
- A units label at the end of the position string.
This current revision comes as I realized adding padding via spaces both is jittery on the hud, and won't look good for all fonts. So the ideal is a matrix of strings that can be repositioned and respaced relatively, via separate draw calls.