Add normal HUD Element checks once override is picked
Eventually as many weapons as possible should get overridden, and once that happens, the default status will be reserved for those that haven't.  The "should the element display" checks may not be used to check for the override, but they certainly should be used to render the override or not.